backend/requirements.txt
alembic==1.13.0
amqp==5.3.1
attrs==25.3.0
babel==2.17.0
bcrypt==4.3.0
billiard==4.2.1
blinker==1.9.0
cachetools==5.5.2
celery==5.5.3
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
coloredlogs==15.0.1
cryptography==41.0.0
dnspython==2.6.0
docopt==0.6.2
email-validator==2.1.0
et_xmlfile==2.0.0
filelock==3.18.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
flatbuffers==25.2.10
fsspec==2025.7.0
google-auth==2.40.3
greenlet==3.2.3
gunicorn==21.2.0
huggingface-hub==0.33.4
humanfriendly==10.0
idna==3.10
imageio==2.37.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
lazy_loader==0.4
llvmlite==0.44.0
Mako==1.3.10
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
num2words==0.5.14
numba==0.61.2
numpy==2.2.6
onnxruntime==1.22.1
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
paramiko==3.5.1
pillow==11.3.0
pip==25.1.1
platformdirs==4.3.8
pluggy==1.6.0
pooch==1.8.2
prompt_toolkit==3.0.51
protobuf==6.31.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
Pygments==2.19.2
PyJWT==2.9.0
PyMatting==1.1.14
PyNaCl==1.5.0
pyreadline3==3.5.4
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
qrcode==8.2
redis==6.2.0
referencing==0.36.2
regex==2023.12.25
rembg==2.0.67
reportlab==4.4.2
requests==2.31.0
rpds-py==0.26.0
rsa==4.9.1
safetensors==0.5.3
scikit-image==0.25.2
scipy==1.16.0
setuptools==80.9.0
six==1.17.0
SQLAlchemy==2.0.25
stripe==8.0.0
sympy==1.14.0
tifffile==2025.6.11
tokenizers==0.21.2
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.53.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.0.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
